# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Article {
  id: string;
  title: string;
  url: string;
  date: string;
  category: string;
  source?: string;
  excerpt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ArticlesSummary {
  total: number;
  byCategory: CategoryCount[];
}

export interface ErrorResponse {
  error: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
}

export interface GeminiMessageInput {
  content: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export interface SummarizeArticlesInput {
  articles: Article[];
}

/**
 * Format of the post to generate
 */
export type GeneratePostInputPostFormat = typeof GeneratePostInputPostFormat[keyof typeof GeneratePostInputPostFormat];


export const GeneratePostInputPostFormat = {
  analysis: 'analysis',
  news: 'news',
  update: 'update',
} as const;

export interface GeneratePostInput {
  article: Article;
  /** Optional custom firm perspective to inject into the prompt */
  firmInsight?: string;
  /** Format of the post to generate */
  postFormat?: GeneratePostInputPostFormat;
}

export type GeneratedPostPostFormat = typeof GeneratedPostPostFormat[keyof typeof GeneratedPostPostFormat];


export const GeneratedPostPostFormat = {
  analysis: 'analysis',
  news: 'news',
  update: 'update',
} as const;

export interface GeneratedPost {
  postFormat?: GeneratedPostPostFormat;
  title: string;
  summaryOfFacts?: string;
  keyTakeaways?: string[];
  firmPerspective?: string;
  articleCategory: string;
  articleDate: string;
  articleUrl: string;
  hashtags?: string[];
  linkedInPost?: string;
  headline?: string;
  summary?: string;
  whyItMatters?: string;
  whatChanged?: string;
  effectiveDate?: string;
  appliesTo?: string;
  actionRequired?: string;
}

export interface ComplianceItem {
  id: string;
  title: string;
  category: string;
  /** YYYY-MM-DD */
  dueDate: string;
  /** Original statutory date if overridden */
  originalDate?: string | null;
  isExtended: boolean;
  isCustom: boolean;
  conditional: boolean;
  conditionNote?: string | null;
  provisional: boolean;
  note?: string | null;
  sourceUrl?: string | null;
  overrideId?: number | null;
}

export interface ComplianceOverrideItem {
  id: number;
  title: string;
  category: string;
  originalDate?: string | null;
  newDate: string;
  note?: string | null;
  sourceUrl?: string | null;
  createdAt: string;
}

export interface CreateComplianceOverrideInput {
  title: string;
  category: string;
  originalDate?: string | null;
  newDate: string;
  note?: string | null;
  sourceUrl?: string | null;
}

export type GetArticlesParams = {
/**
 * Filter articles by publication date in India Standard Time (YYYY-MM-DD format)
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date?: string;
};

export type GetArticlesSummaryParams = {
/**
 * Filter by publication date in India Standard Time (YYYY-MM-DD format)
 * @pattern ^\d{4}-\d{2}-\d{2}$
 */
date?: string;
};

export type GetComplianceCalendarParams = {
/**
 * Calendar year (defaults to current IST year)
 */
year?: number;
/**
 * Calendar month 1-12 (defaults to current IST month)
 * @minimum 1
 * @maximum 12
 */
month?: number;
};

export type DeleteComplianceOverride200 = {
  success: boolean;
};