# lib/api-zod/src/generated/types/complianceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ComplianceItem {
  id: string;
  title: string;
  category: string;
  /** YYYY-MM-DD */
  dueDate: string;
  /** Original statutory date if overridden */
  originalDate?: string | null;
  isExtended: boolean;
  isCustom: boolean;
  conditional: boolean;
  conditionNote?: string | null;
  provisional: boolean;
  note?: string | null;
  sourceUrl?: string | null;
  overrideId?: number | null;
}