# lib/api-zod/src/generated/types/complianceOverrideItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ComplianceOverrideItem {
  id: number;
  title: string;
  category: string;
  originalDate?: string | null;
  newDate: string;
  note?: string | null;
  sourceUrl?: string | null;
  createdAt: string;
}